Add `Mark` and `MarkSet` types for managing cursors etc.
Originally I was using just `(usize, usize)` for marks, and `Vec`s
of those for sets.  But that was already becoming unwieldy even for
really basic code.  This abstracts just a handful of common
operations away to make this easier to reason about.  However,
the internals are (intentionally) left exposed to allow other fiddly
things to be done.